refactor(backend): split files
